Optimize neuron calculations

+++ neurality/neuroninput.py
@@ -1,8 +1,8 @@
 from .neuron import Neuron
 
 class NeuronConnection():
-    def __init__(self, incomingNeuron: Neuron, parentNeuron: Neuron, w: int = 0) -> None:
+    def __init__(self, incomingNeuron: Neuron, parentNeuron: Neuron, w: float = 0.0) -> None:
         self.incomingNeuron = incomingNeuron
         self.parentNeuron = parentNeuron
         self.weight = w
-        self.value = 0
+        self.value = 0.0
